refactor: 테스트 코드 패키지에 맞게 위치 변경
